+++ src/components/fields/FieldInputNumeric.vue
@@ -3,13 +3,15 @@
      ************************************************************************* -->
 
 <template lang="pug">
-.gb-field-input-number
-  field-input-number(
+.gb-field-input-numeric
+  field-input(
     v-bind="$attrs"
     v-on="$listeners"
+    @input="onInput"
+    :value="innerValue"
   )
 
-  .gb-field-input-number__controls
+  .gb-field-input-numeric__controls
 </template>
 
 <!-- *************************************************************************
@@ -23,8 +25,42 @@ import FieldInput from "./FieldInput.vue"
 export default {
   inheritAttrs: false,
 
+  props: {
+    value: {
+      type: Number,
+      default: 0
+    }
+  },
+
   components: {
     FieldInput
+  },
+
+  data: () => ({
+    // --> STATE <--
+
+    innerValue: 0
+  }),
+
+  watch: {
+    value: {
+      immediate: true,
+      handler(value) {
+        // Synchronize inner value with new one
+        this.innerValue = this.value
+      }
+    }
+  },
+
+  methods: {
+    // --> EVENT LISTENERS <--
+
+    onInput(value, name, event) {
+      // Only accept a numeric value
+      if ((event.keyCode >= 48 && event.keyCode <= 57) || (event.keyCode >= 96 && event.keyCode <= 105)) {
+        this.innerValue = value
+      }
+    }
   }
 }
 </script>
@@ -35,5 +71,5 @@ export default {
 
 <style lang="scss">
 // VARIABLES
-$c: ".gb-field-input-number";
+$c: ".gb-field-input-numeric";
 </style>